membuat class baru yaitu farm untuk memindahkan beberapa parameter yang ada di employee yang berisi informasi family
